Replicate an async function a number of times

Similar to base::replicate(), with some differences:

it takes an async function, instead of an expression, and

it always returns a list.

async_replicate(n, task, ..., .limit = Inf)

Arguments

n

Number of replications.

task

Async function to call.

...

Additional arguments to task.

.limit

Number of concurrent async processes to create.

Value

Resolves to a list of the results of the n

task calls.

Examples

# \donttest{
## perform an HTTP request three times, and list the reponse times
do <- function() {
  async_replicate(3,
    function() http_get("https://eu.httpbin.org")$then(function(x) x$times))
}
synchronise(do())
#> [[1]]
#>      redirect    namelookup       connect   pretransfer starttransfer 
#>      0.000000      0.014173      0.085736      0.401632      0.478101 
#>         total 
#>      0.478148 
#> 
#> [[2]]
#>      redirect    namelookup       connect   pretransfer starttransfer 
#>      0.000000      0.014131      0.164929      0.401637      0.954442 
#>         total 
#>      0.954572 
#> 
#> [[3]]
#>      redirect    namelookup       connect   pretransfer starttransfer 
#>      0.000000      0.014104      0.243313      0.475541      0.553833 
#>         total 
#>      0.553922 
#> 
# }