combine {async}	R Documentation

Combine several channels into one.

Description

combine(...) takes any number of promise or channel objects. It awaits each one, and returns a channel object which re-emits every value from its targets, in whatever order they are received.

Usage

combine(...)

Arguments

...
Each argument should be a promise or a channel.

Value

a channel object.

Author(s)

Peter Meilstrup

[Package async version 0.3.2 Index]