Gather all values emitted by iterators and channels.

These functions help collect values from generators, streams or other processes into lists. Note that you can also generate a list of values using run.

as.list.iteror and as.vector.iteror convert iterable objects into vectors of the given mode.

gather takes a channel as argument and returns a promise. All values emitted by the channel will be collected into a vector matching the prototype mode. After the source channel closes, the promise will resolve with the collected vector.

Method as.promise.channel is a synonym for gather.

collect and collector are used in the implementation of the above functions. collect calls the function fn in its argument, supplying a callback of the form function (val, name=NULL). I like to call it emit. While fn is running, it can call emit(x) any number of times. After fn returns, all the values passed to emit are returned in a vector, with optional names.

collector() works similarly to collect() but does not gather values when your inner function returns. Instead, it provides your inner function with two callbacks, one to add a value and the second to extract the value; so you can use that callback to extract values at a later time. For an example of collector usage see the definition of gather.

# S3 method for iteror
as.list(x, ...)

# S3 method for iteror
as.vector(x, mode)

gather(ch, type = list())

# S3 method for channel
as.promise(ch, type = list())

collect(fn, type = list())

collector(fn, type = list())

Arguments

x

An iteror

...

Undocumented.

mode

The mode of the output; taking the same modes as vector.

ch

a channel object.

type

A prototype output vector (similar to the FUN.VALUE argument of vapply) Defaults to list().

fn

A function, which should accept a single argument, here called emit.

Value

as.list.iteror returns a [list].

as.vector.iteror(x, mode) returns a vector of the given mode.

gather(ch, list()) returns a [promise] that eventually resolves with a list. If the channel emits an error, the promise will reject with that error. The partial results will be attached to the error's attr(err, "partialResults").

collect returns a vector of the same mode as type.

Author

Peter Meilstrup

Examples

as.list(iseq(1,10, by=3))
#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 4
#> 
#> [[3]]
#> [1] 7
#> 
#> [[4]]
#> [1] 10
#> 

as.vector(gen(for (i in 1:10) if (i %% 3 != 0) yield(i)), "numeric")
#> [1]  1  2  4  5  7  8 10

ch <- stream(for (i in 1:10) {await(delay(0.1)); if (i %% 3 == 0) yield(i)})
if (FALSE)  ch |> gather(numeric(0)) |> then(\(x)cat(x, "\n")) 

#cumulative sum with collect
cumsum <- function(vec) {
  total <- 0
  collect(type=0, function(emit) {
    for (i in vec) total <- emit(total+i)
  })
}

# `as.list.iteror` is implemented simply with `collect`:
as.list.iteror <- function(it) {
  collect(\(yield) repeat yield(nextOr(it, break)))
}